version 2
This is a more personalized one. I added some files and packages that I'm used to working with.

If you are going to run this, keep in mind that some of those packages require you to add extra configs to settings.py. To stay on top of these configs make a quick look at its documentation on pypi.org.